requirements.txt
pymongo==3.6.1
wordcloud==1.4.1
numpy==1.15.0
jieba==0.39
matplotlib==2.2.2
requests==2.18.4
Pillow==5.3.0
